+++ src/cli/mod.rs
@@ -111,6 +111,7 @@ mod tests {
                     volume: 100,
                     data_notification: Default::default(),
                     favorite_notification: Default::default(),
+                    remote_notifications: Default::default(),
                 },
                 style: StyleType::Custom(ExtraStyles::DraculaDark),
             },